+++ all/rmemo/index.js
@@ -12,20 +12,19 @@ let $ = globalThis.__rmemo__ ??= { q: new Set }
  * @private
  */
 export function memo_(memo_def, ...add_def_a) {
+	let t = []
 	/** @type {memo_T} */
 	let memo = ()=>{
-		if (!('val' in memo)) {
-			memo.f()
-		}
+		'val' in memo || memo.f()
 		if ($.c) {
-			if (!memo.t.includes($.c.s ??= new WeakRef($.c.f))) {
-				memo.t.push($.c.s)
+			if (!t.includes($.c.r ??= new WeakRef($.c))) {
+				t.push($.c.r)
 			}
-			if ($.c.f.l < memo.f.l + 1) $.c.f.l = memo.f.l + 1
+			if ($.c.l < memo.l + 1) $.c.l = memo.l + 1
 			// memo is called by $.c's conditional execution...next change to memo will notify $.c
-			$.c.f.s.push(memo)
+			$.c.t.push(memo)
 			// prevent memo from GC while $.c still has a strong reference
-			if (!$.c.f.t.includes(memo)) $.c.f.t.push(memo)
+			if (!$.c.i.includes(memo)) $.c.i.push(memo)
 		}
 		return memo.val
 	}
@@ -33,9 +32,9 @@ export function memo_(memo_def, ...add_def_a) {
 		get: memo,
 		set: val=>{
 			if (memo.val !== val) {
-				memo.t = memo.t.filter(r=>{
+				t = t.filter(r=>{
 					r = r.deref()
-					if (r?.s.includes(memo)) { // if added by $.c.f.s.push(memo), add to $.q
+					if (r?.t.includes(memo)) { // if added by $.c.t.push(memo), add to $.q
 						$.q.add(r)
 					}
 					return r
@@ -54,7 +53,7 @@ export function memo_(memo_def, ...add_def_a) {
 						continue cur_refresh_loop
 					}
 				}
-				cur_refresh()
+				cur_refresh.f()
 			}
 		},
 	})
@@ -78,7 +77,7 @@ export function memo_(memo_def, ...add_def_a) {
 	memo.f = ()=>{
 		let prev_memo = $.c
 		$.c = memo
-		memo.f.s = [] // reset references in memo_def conditional execution path...see $.c.f.s.push(memo)
+		memo.t = [] // reset references in memo_def conditional execution path...see $.c.t.push(memo)
 		try {
 			memo._ = memo_def(memo)
 		} catch (err) {
@@ -86,10 +85,9 @@ export function memo_(memo_def, ...add_def_a) {
 		}
 		$.c = prev_memo // catch does not throw
 	}
-	memo.f.l = 0
-	memo.f.s = []
-	memo.f.t = []
+	memo.l = 0
 	memo.t = []
+	memo.i = []
 	return memo
 }
 export { memo_ as memosig_ }
@@ -155,8 +153,8 @@ export function memo__bind(fn) {
  */
 export function rmemo__on(rmemo, off_fn) {
 	if (off_fn) rmemo__off__add(rmemo, off_fn)
-	if (rmemo.s?.d) {
-		rmemo.s.deref = rmemo.s.d
+	if (rmemo.r?.d) {
+		rmemo.r.deref = rmemo.r.d
 	}
 	rmemo.f()
 	return rmemo
@@ -167,9 +165,9 @@ export function rmemo__on(rmemo, off_fn) {
  * @returns {rmemo_T}
  */
 export function rmemo__off(rmemo) {
-	if (rmemo.s) {
-		rmemo.s.d ??= rmemo.s.deref
-		rmemo.s.deref = ()=>{
+	if (rmemo.r) {
+		rmemo.r.d ??= rmemo.r.deref
+		rmemo.r.deref = ()=>{
 		}
 	}
 	for (let a_o of rmemo.a ?? []) {